refactoring: package divide
