+++ src/main/java/bowling/README.md
@@ -10,7 +10,7 @@
 
 ## 객체
 
-- 첫 번째 스코어 - 스코어
+첫 번째 스코어 - 스코어
 - 두 번째 스코어
 
 
@@ -42,9 +42,9 @@
 - [ ] 10번째 프레임에서 스트라이크나 스페어가 나오면 한번더 투구를 할 수 있다.
 
 - 볼링 보드
-- [ ] 마지막 프레임에 따라 볼링을 생성한다.
+- [ ] 마지막 프레임에 따라 볼링을 생성한다.-
 
-
+- PlayerBowlingBoard
