+++ src/main/java/bowling/domain/bowlingboard/Player.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.bowlingboard;
 
 public class Player {
