+++ src/main/java/bowling/domain/bowlingboard/ThrowCount.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.bowlingboard;
 
 import java.util.Objects;
