+++ src/main/java/bowling/domain/frame/BowlingFinalFrame.java
@@ -1,5 +1,11 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.BowlingRole;
+import bowling.domain.frame.round.FinalRound;
+import bowling.domain.score.BonusScore;
+import bowling.domain.score.FramePoint;
+import bowling.domain.score.Point;
+import bowling.domain.score.Score;
 import bowling.dto.ScoreDto;
 
 public class BowlingFinalFrame extends BowlingFrame {
