+++ src/main/java/bowling/domain/frame/BowlingNormalFrame.java
@@ -1,5 +1,10 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.BowlingRole;
+import bowling.domain.frame.round.Round;
+import bowling.domain.score.FramePoint;
+import bowling.domain.score.Point;
+import bowling.domain.score.Score;
 import bowling.dto.ScoreDto;
 
 public class BowlingNormalFrame extends BowlingFrame {
