+++ src/main/java/bowling/domain/frame/round/Round.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame.round;
 
 import java.util.Objects;
