+++ src/main/java/bowling/domain/score/FramePoint.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 public class FramePoint {
