+++ src/main/java/bowling/domain/score/Point.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 public class Point {
