+++ src/test/java/bowling/domain/BonusScoreTest.java
@@ -1,5 +1,9 @@
 package bowling.domain;
 
+import bowling.domain.frame.round.NormalRound;
+import bowling.domain.score.BonusScore;
+import bowling.domain.score.Point;
+import bowling.domain.score.Score;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
