+++ src/test/java/bowling/domain/PlayerTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.bowlingboard.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
