+++ src/test/java/bowling/domain/PointTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.score.Point;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
