+++ src/test/java/bowling/domain/ThrowCountTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.bowlingboard.ThrowCount;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
